requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cycler==0.11.0
flatbuffers==2.0
gast==0.5.3
h5py==3.6.0
joblib==1.1.0
keract==4.5.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
protobuf==3.19.4
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.8.0
tensorflow==2.7.0
tensorflow-addons==0.15.0
termcolor==1.1.0
threadpoolctl==3.0.0
typeguard==2.13.3
wrapt==1.13.3